# SlottedArray key preparation