This PR simplifies and speeds up greatly `PrepareKey` `UnPrepareKey` by:

1. extracting hash in a way that is aligned with oddity (3 nibbles for odd, 2 for even) that allows to store always even paths in the array
2. removing one unneeded copy under `Append` when unpreparing the key (all is known and directly written to the output array)
3. reducing if conditions to mostly one optimized switch that results in a tight jump table (` jmp       r8`)

Once the key breaches magic 4 nibbles, optimizations show their true potential

### Benchmarks

A lot of benchmarks testing for oddity, length etc.

#### Before

| Method                | sliceFrom | length | Mean      | Error     | StdDev    | Code Size |
|---------------------- |---------- |------- |----------:|----------:|----------:|----------:|
| Prepare_Key           | 0         | 0      |  1.703 ns | 0.0231 ns | 0.0217 ns |   1,129 B |
| Prepare_Key_UnPrepare | 0         | 0      |  7.377 ns | 0.0798 ns | 0.0747 ns |   2,581 B |
| Prepare_Key           | 0         | 1      |  2.322 ns | 0.0146 ns | 0.0137 ns |   1,029 B |
| Prepare_Key_UnPrepare | 0         | 1      |  7.600 ns | 0.0908 ns | 0.0805 ns |   2,355 B |
| Prepare_Key           | 0         | 2      |  2.810 ns | 0.0448 ns | 0.0419 ns |   1,056 B |
| Prepare_Key_UnPrepare | 0         | 2      |  7.470 ns | 0.0670 ns | 0.0627 ns |   2,382 B |
| Prepare_Key           | 0         | 3      |  3.356 ns | 0.0377 ns | 0.0353 ns |   1,088 B |
| Prepare_Key_UnPrepare | 0         | 3      |  7.681 ns | 0.0840 ns | 0.0786 ns |   2,332 B |
| Prepare_Key           | 0         | 4      |  4.069 ns | 0.0349 ns | 0.0327 ns |   1,123 B |
| Prepare_Key_UnPrepare | 0         | 4      |  7.799 ns | 0.0748 ns | 0.0700 ns |   2,367 B |
| Prepare_Key           | 0         | 6      |  4.259 ns | 0.0424 ns | 0.0397 ns |     861 B |
| Prepare_Key_UnPrepare | 0         | 6      | 13.324 ns | 0.1259 ns | 0.1116 ns |   3,038 B |
| Prepare_Key           | 0         | 32     |  4.268 ns | 0.0626 ns | 0.0586 ns |     861 B |
| Prepare_Key_UnPrepare | 0         | 32     | 27.583 ns | 0.2081 ns | 0.1947 ns |   3,046 B |
| Prepare_Key           | 1         | 1      |  1.757 ns | 0.0235 ns | 0.0220 ns |   1,129 B |
| Prepare_Key_UnPrepare | 1         | 1      |  7.422 ns | 0.0494 ns | 0.0462 ns |   2,581 B |
| Prepare_Key           | 1         | 2      |  2.847 ns | 0.0239 ns | 0.0200 ns |   1,056 B |
| Prepare_Key_UnPrepare | 1         | 2      |  7.676 ns | 0.0545 ns | 0.0510 ns |   2,418 B |
| Prepare_Key           | 1         | 3      |  3.351 ns | 0.0246 ns | 0.0230 ns |   1,088 B |
| Prepare_Key_UnPrepare | 1         | 3      |  7.662 ns | 0.0934 ns | 0.0828 ns |   2,340 B |
| Prepare_Key           | 1         | 4      |  4.085 ns | 0.0436 ns | 0.0408 ns |   1,123 B |
| Prepare_Key_UnPrepare | 1         | 4      |  8.028 ns | 0.0963 ns | 0.0900 ns |   2,365 B |
| Prepare_Key           | 1         | 6      |  4.242 ns | 0.0459 ns | 0.0429 ns |     861 B |
| Prepare_Key_UnPrepare | 1         | 6      | 15.579 ns | 0.1362 ns | 0.1274 ns |   3,105 B |
| Prepare_Key           | 1         | 30     |  4.251 ns | 0.0693 ns | 0.0648 ns |     861 B |
| Prepare_Key_UnPrepare | 1         | 30     | 28.856 ns | 0.3314 ns | 0.3100 ns |   3,132 B |
| Prepare_Key           | 1         | 31     |  4.304 ns | 0.0587 ns | 0.0549 ns |     861 B |
| Prepare_Key_UnPrepare | 1         | 31     | 29.727 ns | 0.1927 ns | 0.1708 ns |   3,205 B |

#### After

| Method                | sliceFrom | length | Mean      | Error     | StdDev    | Code Size |
|---------------------- |---------- |------- |----------:|----------:|----------:|----------:|
| Prepare_Key           | 0         | 0      |  1.839 ns | 0.0192 ns | 0.0179 ns |     728 B |
| Prepare_Key_UnPrepare | 0         | 0      |  7.246 ns | 0.0851 ns | 0.0710 ns |   2,652 B |
| Prepare_Key           | 0         | 1      |  1.554 ns | 0.0278 ns | 0.0260 ns |     687 B |
| Prepare_Key_UnPrepare | 0         | 1      |  7.442 ns | 0.0465 ns | 0.0412 ns |   2,515 B |
| Prepare_Key           | 0         | 2      |  1.536 ns | 0.0308 ns | 0.0288 ns |     690 B |
| Prepare_Key_UnPrepare | 0         | 2      |  7.548 ns | 0.1134 ns | 0.1061 ns |   2,536 B |
| Prepare_Key           | 0         | 3      |  1.606 ns | 0.0261 ns | 0.0244 ns |     690 B |
| Prepare_Key_UnPrepare | 0         | 3      |  7.484 ns | 0.0348 ns | 0.0325 ns |   2,532 B |
| Prepare_Key           | 0         | 4      |  1.591 ns | 0.0240 ns | 0.0201 ns |     689 B |
| Prepare_Key_UnPrepare | 0         | 4      |  7.459 ns | 0.0957 ns | 0.0895 ns |   2,518 B |
| Prepare_Key           | 0         | 6      |  3.318 ns | 0.0266 ns | 0.0222 ns |     719 B |
| Prepare_Key_UnPrepare | 0         | 6      | 12.078 ns | 0.0896 ns | 0.0838 ns |   2,841 B |
| Prepare_Key           | 0         | 32     |  3.375 ns | 0.0649 ns | 0.0607 ns |     719 B |
| Prepare_Key_UnPrepare | 0         | 32     | 12.803 ns | 0.1015 ns | 0.0950 ns |   2,867 B |
| Prepare_Key           | 1         | 1      |  1.544 ns | 0.0259 ns | 0.0229 ns |     687 B |
| Prepare_Key_UnPrepare | 1         | 1      |  7.518 ns | 0.0743 ns | 0.0695 ns |   2,536 B |
| Prepare_Key           | 1         | 2      |  1.666 ns | 0.0275 ns | 0.0258 ns |     690 B |
| Prepare_Key_UnPrepare | 1         | 2      |  7.589 ns | 0.0724 ns | 0.0677 ns |   2,532 B |
| Prepare_Key           | 1         | 3      |  1.652 ns | 0.0338 ns | 0.0316 ns |     689 B |
| Prepare_Key_UnPrepare | 1         | 3      |  7.553 ns | 0.0919 ns | 0.0860 ns |   2,532 B |
| Prepare_Key           | 1         | 4      |  1.776 ns | 0.0170 ns | 0.0142 ns |     689 B |
| Prepare_Key_UnPrepare | 1         | 4      |  7.530 ns | 0.0788 ns | 0.0737 ns |   2,500 B |
| Prepare_Key           | 1         | 6      |  3.058 ns | 0.0378 ns | 0.0353 ns |     684 B |
| Prepare_Key_UnPrepare | 1         | 6      | 10.787 ns | 0.1277 ns | 0.1194 ns |   2,748 B |
| Prepare_Key           | 1         | 30     |  3.031 ns | 0.0208 ns | 0.0194 ns |     684 B |
| Prepare_Key_UnPrepare | 1         | 30     | 11.382 ns | 0.1041 ns | 0.0973 ns |   2,758 B |
| Prepare_Key           | 1         | 31     |  3.035 ns | 0.0273 ns | 0.0255 ns |     684 B |
| Prepare_Key_UnPrepare | 1         | 31     | 11.511 ns | 0.0983 ns | 0.0919 ns |   2,783 B |